[#491] Move StringLiteral proc macro to ffi-macros
