+++ iceoryx2-ffi/ffi-macros/src/lib.rs
@@ -11,9 +11,13 @@
 // SPDX-License-Identifier: Apache-2.0 OR MIT
 
 use proc_macro::TokenStream;
+use proc_macro2::Literal;
 use proc_macro2::TokenTree;
 use quote::{format_ident, quote};
-use syn::{parse_macro_input, punctuated::Punctuated, DeriveInput, LitStr, Meta, Token};
+use syn::{
+    parse_macro_input, punctuated::Punctuated, Data, DeriveInput, Expr, ExprLit, Fields, Lit,
+    LitStr, Meta, Token,
+};
 
 #[proc_macro_attribute]
 pub fn iceoryx2_ffi(args: TokenStream, input: TokenStream) -> TokenStream {
@@ -216,4 +220,130 @@ fn parse_attribute_args(args: TokenStream) -> Args {
     // };
 
     Args { rust_type }
+}
+
+/// Implements the [`iceoryx2_bb_elementary::AsStringLiteral`] trait for enums to provide a string representation of each enum variant.
+///
+/// The string representation can be customized using the `CustomString` attribute, otherwise it will
+/// convert the variant name to lowercase and replace underscores with spaces.
+///
+/// # Example
+/// ```
+/// use iceoryx2_bb_derive_macros::StringLiteral;
+/// use iceoryx2_bb_elementary::AsStringLiteral;
+///
+/// #[derive(StringLiteral)]
+/// enum MyEnum {
+///     #[CustomString = "custom variant one"]
+///     VariantOne,
+///     VariantTwo,
+/// }
+///
+/// let v1 = MyEnum::VariantOne;
+/// assert_eq!(v1.as_str_literal(), "custom variant one\0");
+///
+/// let v2 = MyEnum::VariantTwo;
+/// assert_eq!(v2.as_str_literal(), "variant two\0");
+/// ```
+#[proc_macro_derive(StringLiteral, attributes(CustomString))]
+pub fn string_literal_derive(input: TokenStream) -> TokenStream {
+    let input = parse_macro_input!(input as DeriveInput);
+    let name = &input.ident;
+    let (impl_generics, type_generics, where_clause) = input.generics.split_for_impl();
+
+    // Generate implementation converting enums to a string representation
+    let as_string_literal_impl = match input.data {
+        Data::Enum(ref data_enum) => {
+            let enum_to_string_mapping = data_enum.variants.iter().map(|variant| {
+                let enum_name = &variant.ident;
+                let enum_string_literal = variant
+                    .attrs
+                    .iter()
+                    .find_map(|attr| {
+                        if !attr.path().is_ident("CustomString") {
+                            return None;
+                        }
+                        // Get the value of CustomString as a string literal
+                        match attr.meta.require_name_value() {
+                            Ok(meta) => match &meta.value {
+                                Expr::Lit(ExprLit {
+                                    lit: Lit::Str(lit), ..
+                                }) => {
+                                    let mut value = lit.value();
+                                    value.push('\0');
+                                    Some(Literal::string(&value))
+                                }
+                                _ => None,
+                            },
+                            _ => None,
+                        }
+                    })
+                    .unwrap_or_else(|| {
+                        // If no CustomString, generates default string literal in the form
+                        // MyEnum::MyVariantName => 'my variant name'
+                        let mut enum_string_literal = enum_name
+                            .to_string()
+                            .chars()
+                            .fold(String::new(), |mut acc, c| {
+                                if c.is_uppercase() && !acc.is_empty() {
+                                    acc.push('_');
+                                }
+                                acc.push(c);
+                                acc
+                            })
+                            .chars()
+                            .map(|c| match c {
+                                '_' => ' ',
+                                c => c.to_ascii_lowercase(),
+                            })
+                            .collect::<String>();
+                        enum_string_literal.push('\0');
+                        Literal::string(&enum_string_literal)
+                    });
+
+                // Maps each enum variant to its string representation
+                match &variant.fields {
+                    Fields::Unit => {
+                        quote! {
+                            Self::#enum_name => #enum_string_literal
+                        }
+                    }
+                    Fields::Unnamed(_) => {
+                        quote! {
+                            Self::#enum_name(..) => #enum_string_literal
+                        }
+                    }
+                    Fields::Named(_) => {
+                        quote! {
+                            Self::#enum_name{..} => #enum_string_literal
+                        }
+                    }
+                }
+            });
+
+            // Generate the mapping for the enum variant
+            quote! {
+                fn as_str_literal(&self) -> &'static str {
+                    match self {
+                        #(#enum_to_string_mapping,)*
+                    }
+                }
+            }
+        }
+        _ => {
+            // Does not work for non-enum types
+            let err =
+                syn::Error::new_spanned(&input, "AsStringLiteral can only be derived for enums");
+            return err.to_compile_error().into();
+        }
+    };
+
+    // Implement the trait with the generated implementation
+    let expanded = quote! {
+        impl #impl_generics AsStringLiteral for #name #type_generics #where_clause {
+            #as_string_literal_impl
+        }
+    };
+
+    TokenStream::from(expanded)
 }
