+++ iceoryx2-ffi/ffi/src/api/node.rs
@@ -23,10 +23,10 @@ use iceoryx2::node::{
 };
 use iceoryx2::prelude::*;
 use iceoryx2_bb_container::semantic_string::SemanticString;
-use iceoryx2_bb_derive_macros::StringLiteral;
 use iceoryx2_bb_elementary::static_assert::*;
 use iceoryx2_bb_elementary::AsStringLiteral;
 use iceoryx2_ffi_macros::iceoryx2_ffi;
+use iceoryx2_ffi_macros::StringLiteral;
 
 use core::ffi::{c_char, c_int};
 use core::mem::ManuallyDrop;
